+++ spec/proxy_spec.lua
@@ -52,12 +52,6 @@ describe('Proxy', function()
     local get_upstream
     before_each(function() get_upstream = proxy.get_upstream end)
 
-    it('sets correct upstream port', function()
-      assert.same(443, get_upstream({ api_backend = 'https://example.com' }):port())
-      assert.same(80, get_upstream({ api_backend = 'http://example.com' }):port())
-      assert.same(8080, get_upstream({ api_backend = 'http://example.com:8080' }):port())
-    end)
-
     it("on invalid api_backend return error", function()
       local upstream, err = get_upstream({ api_backend = 'test.com' })
       assert.falsy(upstream)